# jerrinot's improvement: fast path for name len <= 16